==> 01_ETL .ipynb <==
# Importamos las librerías generales
import pandas as pd
import json
import gzip
import numpy as np
import pyarrow as pa
import pyarrow.parquet as pq

# Importamos las bibliotecas de procesamiento de texto y análisis de sentimiento
import nltk
nltk.download('vader_lexicon')
from nltk.sentiment.vader import SentimentIntensityAnalyzer


### Objetivos generales de ETL inicial

- Importar correctamente datos de origen JSON.

- De ser necesario. desanidar datos. 

- Eliminar columnas irrelevantes, manejar valores nulos y duplicados.

- De ser necesario, normalizar tipo de datos.

- Crear la columna "sentiment_analysis" aplicando NLP para el dataframe Reviews.

- Realizar trasforamciones para la correcta utilizacion de los datos por parte de las funciones.

- Guardar el dataset ya procesado en formato Parquet.


# REVIEWS


# Lista para almacenar diccionarios creados a partir de cada línea
lista_user_reviews = []

# Ruta del archivo comprimido en modo lectura
archivo_comprimido = 'datasets/raw_data/user_reviews.json.gz'

# Abrir el archivo comprimido y procesar cada línea
with gzip.open(archivo_comprimido, 'rt', encoding='utf-8') as archivo:
    for linea in archivo:
        try:
            # Utilizar eval para convertir cada línea en un diccionario
            json_data = eval(linea)
            lista_user_reviews.append(json_data)
        except ValueError as e:
            # Manejar errores al intentar evaluar la línea como un diccionario 
            print(f"Error en la línea: {linea}")
            continue

# Crear un DataFrame a partir de la lista de diccionarios
reviews = pd.DataFrame(lista_user_reviews)
# Creamos una copia para trabajar
df_user_reviews = reviews.copy()


df_user_reviews.head() # hay datos anidados en reviews

# Usamos json_normalize para extraer la informacion anindada en la columna 'reviews'
df_user_reviews = pd.json_normalize(
    # Utilizamos la lista creada al principio
    data=lista_user_reviews, # Json_normalize solo admite objetos no serializados para funcionar
    record_path='reviews', # Indicamos la columna con los datos anidados
    meta=['user_id', 'user_url'] # Campos que se quieren corsevar
)

# Hay columnas innecesarias
df_user_reviews.columns


# Eliminamos columnas innecesarias
df_user_reviews.drop(['funny', 'posted', 'last_edited', 'helpful', 'user_url'], axis=1, inplace=True)

df_user_reviews.head(5) # Observamos el resultado

# Observamos si hay valores duplicados y nulos 
print(f"Existen {df_user_reviews.duplicated().sum()} valores duplicados \n")
print(df_user_reviews.isnull().sum())


# Vemos que no hallamos valores faltantes

# Eliminamos duplicados
df_user_reviews = df_user_reviews.drop_duplicates()

# Volvemos a observar duplicados
print(f"Existen {df_user_reviews.duplicated().sum()} valores duplicados")


> Vamos a crear la columna "sentiment_analysis" sobre df_user_reviews.review aplicando NLP.

# Inicializar el analizador de sentimientos
sia = SentimentIntensityAnalyzer()

# Función para asignar valores según la escala de sentimiento
def obtener_puntuacion_sentimiento(texto):
    if pd.isnull(texto) or texto == '':
        return 1  # Devolver neutral si está vacío o es NaN
    elif isinstance(texto, str):
        sentimiento = sia.polarity_scores(texto)
        puntuacion_compuesta = sentimiento['compound']
        if puntuacion_compuesta >= -0.05:
            return 2  # Buena puntuación
        elif puntuacion_compuesta <= -0.05:
            return 0  # Mala puntuación
        else:
            return 1
    else: 
        return 1  # Devolver neutral para valores que no son str

# Convertir la columna 'review' a cadena de texto
df_user_reviews['review'] = df_user_reviews['review'].astype(str)



# Aplicar la función obtener_puntuacion_sentimiento a la columna 'review'
df_user_reviews['sentiment_analysis'] = df_user_reviews['review'].apply(obtener_puntuacion_sentimiento)



# Eliminar la columna 'review' ya que no la necesitamos con 'sentiment_analysis'
df_user_reviews = df_user_reviews.drop(columns=['review'], axis=1)

# Vemos como queda el df
df_user_reviews

# ITEMS

# Lista para almacenar diccionarios creados a partir de cada línea
lista_users_items = []

# Ruta del archivo comprimido en modo lectura
archivo_comprimido_users_items = 'datasets/raw_data/users_items.json.gz'

# Abrir el archivo comprimido y procesar cada línea
with gzip.open(archivo_comprimido_users_items, 'rt', encoding='utf-8') as archivo_users_items:
    for linea in archivo_users_items:
        try:
            # Utilizar eval para convertir cada línea en un diccionario
            json_data = eval(linea)
            lista_users_items.append(json_data)
        except ValueError as e:
            # Manejar errores al intentar evaluar la línea como un diccionario 
            print(f"Error en la línea: {linea}")
            continue

# Crear un DataFrame a partir de la lista de diccionarios
items = pd.DataFrame(lista_users_items)
# Creamos una copia para trabajar
df_user_items = items.copy()

df_user_items.head() # nuevamente, tenemos datos anidados en items

# Usamos json_normalize para extraer la informacion anindada en la columna 'items'

df_user_items = pd.json_normalize(
    data=lista_users_items, # datos de origen en formato no serializado
    record_path='items', # ruta de datos anidados
    meta=['user_id' ] # campos que se quieren corsevar
)

# Vemos que tenemos columnas inecesarias
df_user_items.columns

# Eliminamos columnas innecesarias
df_user_items.drop(['playtime_2weeks', 'item_name'], axis=1, inplace=True)

# Observamos el resultado
df_user_items.head(5)

# Observamos si hay duplicados y nulos 
print(f"Existen {df_user_items.duplicated().sum()} valores duplicados\n")
print(df_user_items.isnull().sum())

# Eliminamos datos duplicados
df_user_items = df_user_items.drop_duplicates()

print(f"Existen {df_user_items.duplicated().sum()} valores duplicados")


# Eliminar filas donde playtime_forever sea igual a 0
df_user_items = df_user_items[df_user_items['playtime_forever'] != 0]

# GAMES

file_path = r'datasets\raw_data\steam_games.json.gz'

with gzip.open(file_path, 'rt', encoding='utf-8') as file:
    data = [json.loads(line) for line in file]

# Convertir la lista de objetos JSON en un DataFrame
games = pd.DataFrame(data)

# Crear una copia para trabajar
df_steam_games = games.copy()


df_steam_games.head()

# Vemos que tenemos que lidear con una cantidad de NaN muy alta
df_steam_games.isnull().sum()

# Imputar semejante cantidad de fantantes es imposible
df_steam_games.dropna(inplace=True) # eliminamos faltantes y asumimos la utilidad de los datos restantes

# Vemos las columnas para descartar que no son pertinentes
df_steam_games.columns

# Vamos a quitar las columnas innecesarias
df_steam_games = df_steam_games.drop(["publisher","app_name","url","reviews_url",
    "specs","early_access"], axis=1)

# El id corresponde a item_id en df_users_items, por tanto homologamos al nombre de dicho df
df_steam_games = df_steam_games.rename(columns={'id': 'item_id'})


# Quitamos los corchetes de genres para que la funcion pueda consumir la informacion correctamente

# Lista de columnas a procesar
cols = ['genres']

# Itera a través de las columnas
for i in cols:
    # Convierte los valores de la columna en cadenas de texto
    df_steam_games[i] = df_steam_games[i].astype(str)
    
    # Elimina los corchetes '[' ']' de los valores de la columna
    df_steam_games[i] = df_steam_games[i].str.replace('[', '').str.replace(']', '')


df_steam_games

# Creamos la columna release_year para trabajar unicamente con el año
df_steam_games["release_year"] = df_steam_games["release_date"]

# Eliminamos la columna original 
del df_steam_games["release_date"]


# Convertir la columna "release_date" al formato de fecha, permitiendo que Pandas infiera el formato
df_steam_games["release_year"] = pd.to_datetime(df_steam_games["release_year"], format='mixed', errors='coerce')

#  Colocar solo el año de la fecha en "release_date" 
df_steam_games["release_year"] = df_steam_games["release_year"].dt.year

# El parametro errors='coerce' convierte en nulos los datos que no se pueden inferir.

# Contamos los nulos resultantes
df_steam_games["release_year"].isnull().sum()

# Los eliminamos ya que no representan una cantidad significativa
df_steam_games.dropna(inplace=True) 


# Cambiamos los valores de "release_date" a tipo entero
df_steam_games["release_year"] = df_steam_games["release_year"].astype(int)

# Vamos a pasar los str que refieran a juesgos gratuitos a 0.00 
# y luego pasar todos los precios a float

# Función para convertir los valores de la columna "price"
def convertir_precio(valor):
    try:
        # Intentar convertir a flotante
        return float(valor)
    except ValueError:
        # Si no se puede convertir, verificar si es "Free To Play"
        if "Free" in valor:
            return 0.0
        else:
            # Si no es ninguno de los casos anteriores, asignar 0.0
            return 0.0

# Aplicar la función a la columna "price" del DataFrame
df_steam_games['price'] = df_steam_games['price'].apply(convertir_precio)


df_steam_games.head()

# Finalemente exportamos todos los datos limpios

# Convertir los DataFrame  Pandas a objetos de tabla de Arrow
tabla_arrow1 = pa.Table.from_pandas(df_steam_games)
tabla_arrow2 = pa.Table.from_pandas(df_user_reviews)
tabla_arrow3 = pa.Table.from_pandas(df_user_items)

# Especificar el nombre de los archivos Parquet
archivo_parquet1 = 'datasets/processed_data/steam_games.parquet'
archivo_parquet2 = 'datasets/processed_data/user_reviews.parquet'
archivo_parquet3 = 'datasets/processed_data/user_items.parquet'

# Escribir las tablas Arrow en formato Parquet
pq.write_table(tabla_arrow1, archivo_parquet1)
pq.write_table(tabla_arrow2, archivo_parquet2)
pq.write_table(tabla_arrow3, archivo_parquet3)


> ## Resultado final:

**df_steam_games:** Contiene información sobre juegos de Steam, incluyendo nombre, item_id, año de lanzamiento, precio,  desarrollador y género.

**df_user_items:** Registra el tiempo de juego de los usuarios para cada item.

**df_user_reviews:** Contiene información sobre las reseñas de los usuarios, incluyendo item_id, user_id, recomendación (True/False) y análisis del sentimiento (1=negativo,0=neutro,2=positivo).
