==> 04_ML.ipynb <==
import pandas as pd
import pyarrow.parquet as pq
import pyarrow as pa
import pandas as pd
import numpy as np

import nltk

from sklearn.metrics.pairwise import cosine_similarity
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.metrics.pairwise import linear_kernel
from sklearn.preprocessing import MinMaxScaler

from surprise import Dataset, Reader
from surprise import SVD
from surprise.model_selection import GridSearchCV

#### Sistema de recomendación item-item:

- **def recomendacion_juego( item_id ):**

> Ingresando el id de producto, deberíamos recibir una lista con 5 juegos recomendados similares al ingresado.


# Ruta del archivo .parquet
file_path = r'datasets\processed_data\steam_games.parquet'

# Leer el archivo .parquet en un DataFrame
df = pq.read_table(file_path).to_pandas()


# Nos quedamos con las columnas de interes
df = df[['item_id', 'genres']]
# quitamos duplicados y reiniciamos el índice
df = df.drop_duplicates()
df = df.reset_index(drop=True)
df.shape


# Importamos nuestro modelo
tf = TfidfVectorizer()

# calculamos los features (en genres) para cada ítem_id 
tfidf_matrix = tf.fit_transform(df['genres'])

# calculamos las similitudes entre todos los documentos
cosine_similarities = linear_kernel(tfidf_matrix, tfidf_matrix)

def recommend_items(item_id, cosine_similarities, items, n=5):
  # Obtener el índice del elemento actual
  index = np.where(items == item_id)[0][0]

  # Obtener las similitudes del elemento actual con todos los demás elementos
  similar_items = list(enumerate(cosine_similarities[index]))

  # Ordenar las similitudes en orden descendente
  similar_items = sorted(similar_items, key=lambda x: x[1], reverse=True)[1:n+1]

  # Obtener los índices de los elementos más similares
  item_indices = [i[0] for i in similar_items]

  # Devolver los IDs de los elementos más similares
  return items[item_indices]

  # Obtener recomendaciones para un elemento específico
item_id = '252490'
recommendations = recommend_items(item_id, cosine_similarities, df['item_id'])

# Imprimir las recomendaciones
print('Recommendations for item {}:'.format(item_id))
for item in recommendations:
  print(item)

def recommend_items(item_id, cosine_similarities, items, n=5):
    # Obtener el índice del elemento actual
    index = np.where(items == item_id)[0][0]

    # Obtener las similitudes del elemento actual con todos los demás elementos
    similar_items = list(enumerate(cosine_similarities[index]))

    # Ordenar las similitudes en orden descendente
    similar_items = sorted(similar_items, key=lambda x: x[1], reverse=True)[1:n+1]

    # Obtener los índices de los elementos más similares
    item_indices = [i[0] for i in similar_items]

    # Devolver los IDs de los elementos más similares
    return items[item_indices]

# Lista para almacenar las predicciones
all_predictions = []

# Iterar sobre todos los item_id
for item_id in df['item_id'].unique():
    # Obtener las recomendaciones para el item_id actual
    recommendations = recommend_items(item_id, cosine_similarities, df['item_id'])
    # Guardar los resultados en la lista
    all_predictions.append({"item_id": item_id, "top_5_item_id": recommendations})

# Convertir la lista de diccionarios en un DataFrame
predictions_df = pd.DataFrame(all_predictions)



def recommend_items(item_id, cosine_similarities, items, n=5):
    # Obtener el índice del elemento actual
    index = np.where(items == item_id)[0][0]

    # Obtener las similitudes del elemento actual con todos los demás elementos
    similar_items = list(enumerate(cosine_similarities[index]))

    # Ordenar las similitudes en orden descendente
    similar_items = sorted(similar_items, key=lambda x: x[1], reverse=True)[1:n+1]

    # Obtener los índices de los elementos más similares
    item_indices = [i[0] for i in similar_items]

    # Devolver los IDs de los elementos más similares
    return items[item_indices]

# Lista para almacenar las predicciones
all_predictions = []

# Iterar sobre todos los item_id
for item_id in df['item_id'].unique():
    # Obtener las recomendaciones para el item_id actual
    recommendations = recommend_items(item_id, cosine_similarities, df['item_id'])
    # Guardar los resultados en la lista
    all_predictions.append({"item_id": item_id, "top_5_item_id": ' '.join(map(str, recommendations))})


# Exportamos para trabajar los dataos resultantes de las predicciones
item_to_item = pd.DataFrame(all_predictions)
item_to_item.to_parquet('datasets/processed_data/item_to_item.parquet')


# Ruta al archivo Parquet
file_path = 'datasets/processed_data/data_for_ML/item_to_item.parquet'

# Cargar el archivo Parquet en un DataFrame utilizando PyArrow
table = pq.read_table(file_path)
df = table.to_pandas()


# Reemplazar espacios por comas en la columna top_5_item_id
df['top_5_item_id'] = df['top_5_item_id'].str.replace(' ', ',')

df

# Creamos una funcion auxiliar para obtener los primeros 5 item_id
def get_top_5_item_id(item_id):
    row = df[df['item_id'] == item_id]
    if len(row) == 0:
        return None  # Si el item_id no está en el DataFrame, retorna None
    
    top_5 = row['top_5_item_id'].iloc[0].split()[:5]  # Obtiene los primeros 5 item_id de la fila
    return top_5

get_top_5_item_id("761140")


# Nos conviene utilizar un df para obtener el title del item_id correspondiente

# Ruta de los archivos Parquet
games_parquet = 'datasets/processed_data/steam_games.parquet'
# Utiliza PyArrow para leer el archivo Parquet
tabla_games = pq.read_table(games_parquet)
# Leer las tablas en DataFrames de Pandas
df_steam_games = tabla_games.to_pandas()

title_finder = df_steam_games[['item_id', 'title']]


title_finder


def get_top_5_item_id_with_titles(item_id):
    row = df[df['item_id'] == item_id]
    if len(row) == 0:
        return None  # Si el item_id no está en el DataFrame, retorna None
    
    top_5_ids = row['top_5_item_id'].iloc[0].split(",")[:5]  # Divide los top_5_item_id y obtiene los primeros 5
    titles = []
    for id in top_5_ids:
        id = id.strip()  # Elimina espacios en blanco alrededor del item_id
        title_row = title_finder[title_finder['item_id'] == id]['title']
        if not title_row.empty:
            titles.append((id, title_row.iloc[0]))
    return titles


get_top_5_item_id_with_titles("761140")

# Finalmente exportamos nuestros datos a utilizar para la API 
title_finder.to_parquet('datasets/processed_data/data_for_funciones/title_finder.parquet')
df.to_parquet('datasets/processed_data/data_for_funciones/recomendacion_juego.parquet')

# Lo unico que nos queda es adaptar la funcion a la API ...

#### Sistema de recomendación user-item:

- **def recomendacion_usuario( user_id):**

> Ingresando el id de un usuario, deberíamos recibir una lista con 5 juegos recomendados para dicho usuario.

# Especifica la ruta del archivo Parquet
ruta_archivo_parquet = 'datasets/processed_data/data_for_ML/ml_user_items.parquet'

# Lee el archivo Parquet en una tabla de PyArrow
tabla_parquet = pq.read_table(ruta_archivo_parquet)

# Convierte la tabla a un DataFrame de pandas si es necesario
ml_user_items = tabla_parquet.to_pandas()


# Normalizar playtime_forever en un rango de 0 a 1
scaler = MinMaxScaler()
# Crear una nueva columna para dicha normalización
ml_user_items['normalized_playtime'] = scaler.fit_transform(ml_user_items[['playtime_forever']])

# Eliminar las columnas playtime_forever
ml_user_items = ml_user_items.drop(['playtime_forever'], axis=1)

# elimnar duplicados
ml_user_items = ml_user_items.drop_duplicates(subset=['item_id', 'user_id'])



ml_user_items 

# Especificamos el valor mínimo y el valor máximo de los tiempos y el dataset a utilizar
reader = Reader(rating_scale=(ml_user_items["normalized_playtime"].min(),ml_user_items["normalized_playtime"].max()))

# El objeto Dataset de Surprise nos permite leer datos 
dataset = Dataset.load_from_df(ml_user_items,reader)

# Empleamos GridSearchCV con el algoritmo SVD y seteamos el parámetro refit a True con measures = ["rmse","fcp"]
param_grid = {'n_epochs': [5, 10], 'lr_all': [0.002, 0.005], 'reg_all': [0.4, 0.6]}

gs = GridSearchCV(SVD, param_grid, measures=['fcp',"rmse"], cv=3, refit=True)

# Entrenamos el modelo
gs.fit(dataset)

# Mejor combinacion de parametros
gs.best_params

# Mejor Score
gs.best_score

# Guardamos el modelo con mayor fcp
best_model = gs.best_estimator["fcp"]


# Prediccion para user_id : syates,  item_id : 1002 (probamos el modelo)
pred = best_model.predict("syates", "1002")
pred.est

# Definir el user_id para el cual queremos hacer recomendaciones
user_id = "syates"

# Obtener todos los item_id en el dataset
all_item_ids = ml_user_items["item_id"].unique()

# Obtener las predicciones para todos los item_id y seleccionar los 5 con la puntuación más alta
predictions = [(item_id, best_model.predict(user_id, item_id).est) for item_id in all_item_ids]
top_recommendations = sorted(predictions, key=lambda x: x[1], reverse=True)[:5]

# Imprimir las recomendaciones
print(f"Recomendaciones para el usuario {user_id}:")
for item_id, score in top_recommendations:
    print(f"Item ID: {item_id}, Score: {score}")



# Lista para almacenar las predicciones
all_predictions = []

# Iterar sobre todos los user_id
for user_id in ml_user_items["user_id"].unique():
    # Calcular las predicciones para todos los item_id
    predictions = [(item_id, best_model.predict(user_id, item_id).est) for item_id in all_item_ids]
    # Seleccionar los 5 item_id con las puntuaciones más altas
    top_predictions = sorted(predictions, key=lambda x: x[1], reverse=True)[:5]
    # Guardar los resultados en la lista
    all_predictions.append({"user_id": user_id, "top_5_item_id": [item_id for item_id, _ in top_predictions]})

# Convertir la lista de diccionarios en un DataFrame
predictions_df = pd.DataFrame(all_predictions)



predictions_df

def user_recomend(user_id):
    row = predictions_df[predictions_df['user_id'] == user_id]
    if not row.empty:
        return row['top_5_item_id'].iloc[0]
    else:
        return "User ID not found"


# Ejemplo de uso:
user_recomend("syates")

# Abrimos el title_finder
ruta_archivo_parquet = 'datasets/processed_data/data_for_funciones/title_finder.parquet'
tabla_parquet = pq.read_table(ruta_archivo_parquet)
title_finder = tabla_parquet.to_pandas()


def user_recommend_with_titles(user_id):
    row = predictions_df[predictions_df['user_id'] == user_id]
    if not row.empty:
        top_5_ids = row['top_5_item_id'].iloc[0][:5]  # Obtenemos los primeros 5 elementos de la lista
        titles = []
        for id in top_5_ids:
            id = id.strip()  # Elimina espacios en blanco alrededor del item_id
            title_row = title_finder[title_finder['item_id'] == id]['title']
            if not title_row.empty:
                titles.append(title_row.iloc[0])
        return titles
    else:
        return "User ID not found"


user_recommend_with_titles("syates")

# exportamos los datos para la funcion en la API
predictions_df.to_parquet('datasets/processed_data/data_for_funciones/recomendacion_usuario.parquet')
predictions_df.to_csv('datasets/processed_data/data_for_funciones/recomendacion_usuario.csv')
